feat(ResultCard) add result card component
