+++ src/views/Flipper.vue
@@ -1,24 +1,7 @@
 <template>
   <div>
         <v-img :src='`${resourceURL}/static/image/pool_image/${pool}.png`'></v-img>
-
-        <v-img
-            class="mt-2"
-            gradient="to top right, rgba(100,115,201,.33), rgba(25,32,72,.7)"
-            :src='`https://tw.hicdn.beanfun.com/beanfun/WebImage/1606509847551.png`'
-            height="300px"
-            width="440px"
-        >
-          <v-container class="mt-2">
-            <v-row class="justify-center mb-3" v-for="(row, index) in [3, 4, 3]" :key="`row_${pool}_${row}`">
-              <character-icon v-for="n in row" :key="n"
-                              :attri="result[index][n]? result[index][n].attri: ''"
-                              :name="result[index][n]? result[index][n].name: ''"
-                              :id="result[index][n]? result[index][n].id: ''"
-                              :rarity="result[index][n]?result[index][n].rarity:''"/>
-            </v-row>
-          </v-container>
-        </v-img>
+        <result-card :character-list="result"></result-card>
 
         <v-img :src='require("..\\assets\\bg.jpg")'>
           <v-container class="justify-center mt-3">
@@ -67,11 +50,11 @@
 <script>
 import API from '../plugins/api';
 import RollButton from '@/components/RollButton';
-import CharacterIcon from '@/components/CharacterIcon';
+import ResultCard from '@/components/ResultCard';
 
 export default {
   name: 'Flipper',
-  components: { CharacterIcon, RollButton },
+  components: { ResultCard, RollButton },
   props: {
     pool: {
       type: String,
@@ -82,7 +65,7 @@ export default {
     return {
       resourceURL: API.resourceURL,
       isRoll: false,
-      result: [[], [], []],
+      result: [],
       resultReport: [0, 0, 0],
       total: 0
     };
@@ -90,32 +73,15 @@ export default {
   methods: {
     roll () {
       API.roll(this.pool).then((rs) => {
-
-        let rollDate = [[], [], []];
         let report = { '3': 0, '4': 0, '5': 0 };
 
-        rs.data.forEach((i, index) => {
+        rs.data.forEach(i => {
           report[i.rarity]++;
-
-          switch (true) {
-            case index < 3:
-              rollDate[0][index + 1] = i;
-              break;
-            case index < 7:
-              rollDate[1][index - 2] = i;
-              break;
-            case index < 10:
-              rollDate[2][index - 6] = i;
-              break;
-            default:
-              //TODO Error;
-              break;
-          }
         });
 
         this.total = rs.total;
         this.resultReport = [report['5'], report['4'], report['3']];
-        this.result = rollDate;
+        this.result = rs.data;
         this.isRoll = true;
       });
     }
